hxxx_nal: use size_t for the move values
They seem to represent a size for data in memory (size in block allocation) and
offset for a buffer memmove().

+++ modules/packetizer/hxxx_nal.c
@@ -49,7 +49,7 @@ block_t *hxxx_AnnexB_to_xVC( block_t *p_block, uint8_t i_nal_length_size )
     {
         const uint8_t *p; /* start of prefixed nal */
         uint8_t  prefix; /* startcode length */
-        off_t    move; /* move offset */
+        size_t   move; /* move offset */
     } *p_list = NULL;
 
     if(!p_block->i_buffer || p_block->p_buffer[0])
@@ -62,7 +62,7 @@ block_t *hxxx_AnnexB_to_xVC( block_t *p_block, uint8_t i_nal_length_size )
     const uint8_t *p_buf = p_block->p_buffer;
     const uint8_t *p_end = &p_block->p_buffer[p_block->i_buffer];
     unsigned i_bitflow = 0;
-    off_t i_move = 0;
+    size_t i_move = 0;
     while( p_buf != p_end )
     {
         i_bitflow <<= 1;
@@ -82,7 +82,7 @@ block_t *hxxx_AnnexB_to_xVC( block_t *p_block, uint8_t i_nal_length_size )
                 p_list[i_nalcount].p = &p_buf[-2];
                 p_list[i_nalcount].prefix = 3;
             }
-            i_move += (off_t) i_nal_length_size - p_list[i_nalcount].prefix;
+            i_move += (size_t) i_nal_length_size - p_list[i_nalcount].prefix;
             p_list[i_nalcount++].move = i_move;
 
             /* Check and realloc our list */
@@ -147,7 +147,7 @@ block_t *hxxx_AnnexB_to_xVC( block_t *p_block, uint8_t i_nal_length_size )
     {
         const uint8_t *p_readstart = p_list[i - 1].p;
         uint32_t i_payload = p_sourceend - p_readstart - p_list[i - 1].prefix;
-        off_t offset = p_list[i - 1].p - p_source + p_list[i - 1].prefix + p_list[i - 1].move;
+        size_t offset = p_list[i - 1].p - p_source + p_list[i - 1].prefix + p_list[i - 1].move;
 //        printf(" move offset %ld, length = %ld  prefix %ld move %ld\n", p_readstart - p_source, i_payload, p_list[i - 1].prefix, p_list[i-1].move);
 
         /* move in same / copy between buffers */
